feat: correctly receive image
